backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.1
beanie==1.13.1
click==8.1.7
dnspython==2.7.0
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.98.0
h11==0.14.0
idna==3.10
motor==3.6.0
multidict==6.1.0
passlib==1.7.4
propcache==0.2.0
pyasn1==0.6.1
pydantic==1.10.19
pydantic_core==2.23.4
PyJWT==2.10.1
pymongo==4.9.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.27.0
toml==0.10.2
typing_extensions==4.12.2
uvicorn==0.32.0
yarl==1.17.2
